Use standard migration setup for this engine
