--- lib/generators/active_record/templates/add_event_source_to_best_boy_events_table.rb
@@ -1,11 +1,0 @@
-class AddEventSourceToBestBoyEventsTable < ActiveRecord::Migration
-  def self.up
-    add_column :best_boy_events, :event_source, :string
-    add_index :best_boy_events, :event_source
-  end
-
-  def self.down
-    remove_index :best_boy_events, :event_source
-    remove_column :best_boy_events, :event_source
-  end
-end
