+++ lib/generators/best_boy/templates/config/initializers/best_boy.rb
