--- spec/dummy/config/initializers/best_boy.rb
@@ -1,13 +1,0 @@
-BestBoy.setup do |config|
-  # Define base Controller (default ApplicationController)
-  config.base_controller = "ApplicationController"
-
-  # Define before_filter for using before_filters (default = nil)
-  #config.before_filter = nil
-
-  # Define skip_before_filter for skipping before_filters (default = nil)
-  #config.skip_before_filter = nil
-
-  # Define custom redirect path as string p.e. "/admin" (default = nil)
-  #config.custom_redirect = nil
-end
